Fix non-interleaved sample
